requirements.txt
discord-components==1.1.4
discord.py==1.7.1
mysql-connector-python==8.0.20